crazzy project/real_time_gpu_filters/requirements.txt
numpy>=1.21.0
opencv-python>=4.8.0
# pycuda>=2022.1  # Comment out problematic version
PyQt5>=5.15.0
pillow>=9.0.0
matplotlib>=3.5.0
# CuPy (select the build matching your CUDA version, e.g., CUDA 12.x shown below)
cupy-cuda12x>=12.0.0
# Additional GPU libraries as alternatives
numba>=0.56.0
psutil>=5.9.0
pynvml>=11.5.0
